# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApplicationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ApplicationPage",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Stable, for your code",
                "type": "string"
              },
              "details": {
                "description": "Depends on code",
                "type": "object"
              },
              "message": {
                "description": "For people; may change",
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A sign-in token from the company identity provider, or from /dev/token in dev.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Talent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/applications": {
      "get": {
        "description": "Applications in your scope, newest first: your own, or all of them for a TA lead.",
        "operationId": "list_applications_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "next_cursor from the last page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "next_cursor from the last page",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "requisition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requisition Id"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z][a-z_]{0,39}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "description": "Follows the candidate: false leaves out archived candidates' applications",
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Follows the candidate: false leaves out archived candidates' applications",
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Applications",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "description": "Attaches a candidate to a requisition. It starts at the first stage, as a transition.",
        "operationId": "create_application_v1_applications_post",
        "parameters": [
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}": {
      "get": {
        "description": "One application: its stage now, its owner, and the transitions allowed from here.",
        "operationId": "get_application_v1_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}/reversal": {
      "post": {
        "description": "Reverses a rejection, with a reason and your name (BR-406). The rejected application stays\nrejected; a new application, linked to it, starts again at the first stage.",
        "operationId": "reverse_rejection_v1_applications__application_id__reversal_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReversalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reverse Rejection",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}/transitions": {
      "get": {
        "description": "Every transition, oldest first: from, to, who and when.",
        "operationId": "transition_history_v1_applications__application_id__transitions_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transition History",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "description": "Moves an application. `from_stage` must be its stage now; a rejection needs a listed\nreason (BR-404); you are recorded as the person who moved it (BR-403).",
        "operationId": "create_transition_v1_applications__application_id__transitions_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Transition",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/v1/reference/stages": {
      "get": {
        "description": "The stage list in force: stages in order and the transitions allowed between them.",
        "operationId": "stages_v1_reference_stages_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stages",
        "tags": [
          "reference"
        ]
      }
    },
...
  }
}